Add sbom files to the flare
Since SBOM is linked to an integration the collector is in charge of
collecting the SBOM file.
