+++ comp/core/flare/flare.go
@@ -183,7 +183,6 @@ func (f *flare) Create(pdata ProfileData, ipcError error) (string, error) {
 		},
 		f.collectLogsFiles,
 		f.collectConfigFiles,
-		f.collectHostSBOM,
 	)
 
 	for _, p := range providers {
